+++ src/compas_rhino/__init__.py
@@ -23,11 +23,8 @@ compas_rhino
 from __future__ import absolute_import
 
 import os
-
 import compas._os
-
 from .utilities import *
-from . import utilities
 
 
 __version__ = '0.7.1'