+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `compas_rhino.geometry.RhinoGeometry` to the docs.
+
 ### Changed
+
+- Fixed missing import in `compas_rhino.geometry`.
+- Removed circular imports from `compas_rhino.geometry`.
 
 ### Removed